Component and styles moved into own files
